Add all necessary files
