+++ bin/ead
@@ -1,4 +1,4 @@
 #!/usr/bin/env ruby
-require "hola"
+require 'ead'
 
 start(ARGV)
