+++ lib/attribute.rb
@@ -1,0 +1,12 @@
+class Attribute
+  attr_accessor :name, :type
+
+  def initialize(name, type)
+    @name = name
+    @type = type
+  end
+
+  def add_to(command)
+    command << " #{name}:#{type}"
+  end
+end
